requirements_tests.txt
hypothesis
pytest
pytest-django
pytest-xdist
pytest-cov
